:construction: refactor and remove duplicate code
