--- plugins/addicted/src/interop/substance/quality.rs
@@ -1,28 +1,0 @@
-use red4ext_rs::prelude::NativeRepr;
-
-#[derive(Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[repr(u32)]
-#[allow(dead_code)]
-pub enum Quality {
-    #[default]
-    Common = 0,
-    CommonPlus = 1,
-    Epic = 2,
-    EpicPlus = 3,
-    Iconic = 4,
-    Legendary = 5,
-    LegendaryPlus = 6,
-    LegendaryPlusPlus = 7,
-    Random = 8,
-    Rare = 9,
-    RarePlus = 10,
-    Uncommon = 11,
-    UncommonPlus = 12,
-    Count = 13,
-    Invalid = 14,
-}
-
-unsafe impl NativeRepr for Quality {
-    const NAME: &'static str = "Quality";
-    const NATIVE_NAME: &'static str = "gamedataQuality";
-}
